# docs/gpt-action-openapi.yaml
openapi: 3.1.0
info:
  title: Timmeny-ToDo-OS API
  version: 0.2.0
  description: Create, read, and organize todo items on the configured Timmeny-ToDo-OS Monday.com boards.
servers:
  - url: https://timmeny-os-production.up.railway.app
paths:
  /todos:
    get:
      operationId: listTodos
      summary: List todo items
      description: Lists todo items from the regular To Do List board, the GS Initiatives & Action Items board, or both, including Action Group, Action Date, and Action metadata when present.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: list
          in: query
          required: false
          schema:
            type: string
            enum:
              - all
              - todo
              - gs
            default: all
          description: Use all for both boards, todo for regular todos, or gs for GS action items.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
          description: Maximum number of items to read from each selected board.
      responses:
        "200":
          description: Todo items read successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoListResponse"
        "401":
          description: Missing or invalid API key.
        "422":
          description: Invalid query parameters.
        "500":
          description: Missing server configuration.
        "502":
          description: Monday.com request failed.
        "504":
          description: Monday.com request timed out.
    post:
      operationId: createTodo
      summary: Create a todo item
      description: Creates a todo item in either the regular To Do List board or the GS Initiatives & Action Items board. Optional action metadata can set Action Group, Action Date, and Action at creation time.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoCreateRequest"
      responses:
        "200":
          description: Todo item created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoCreateResponse"
        "401":
          description: Missing or invalid API key.
        "422":
          description: Invalid request body.
        "500":
          description: Missing server configuration.
        "502":
          description: Monday.com request failed.
        "504":
          description: Monday.com request timed out.
  /todos/{item_id}/action-metadata:
    patch:
      operationId: updateTodoActionMetadata
      summary: Update todo action metadata
      description: Updates the Action Group, Action Date, and/or Action columns for an existing todo item. Use this after reviewing todos and choosing better thematic groupings or identifying missing decisions.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: string
          description: Monday.com item id to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoUpdateActionMetadataRequest"
      responses:
        "200":
          description: Todo action metadata updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoUpdateActionMetadataResponse"
        "401":
          description: Missing or invalid API key.
        "422":
          description: Invalid request body.
        "500":
          description: Missing server configuration.
        "502":
          description: Monday.com request failed.
        "504":
          description: Monday.com request timed out.
  /todos/bulk-action-metadata:
    post:
      operationId: bulkUpdateTodoActionMetadata
      summary: Bulk update todo action metadata
      description: Updates Action Group, Action Date, and/or Action for multiple existing todo items in one request. Use this when applying grouping changes to more than one item.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoBulkActionMetadataRequest"
      responses:
        "200":
          description: Bulk update completed. Check each result for per-item success or failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoBulkActionMetadataResponse"
        "401":
          description: Missing or invalid API key.
        "422":
          description: Invalid request body.
        "500":
          description: Missing server configuration.
        "502":
          description: Monday.com request failed.
        "504":
          description: Monday.com request timed out.
components:
  schemas:
    TodoListName:
      type: string
      enum:
        - todo
        - gs
      description: Use todo for the regular To Do List board and gs for GS Initiatives & Action Items.
    TodoActionMetadata:
      type: object
      additionalProperties: false
      properties:
        action_group:
          type:
            - string
            - "null"
          maxLength: 255
          description: The thematic grouping to write into the Monday.com Action Group text column.
        action_date:
          type:
            - string
            - "null"
          format: date
          description: The date to write into the Monday.com Action Date column, formatted YYYY-MM-DD.
        action:
          type:
            - string
            - "null"
          maxLength: 255
          description: The action type label to write into the Monday.com Action dropdown column, such as Decision.
    TodoCreateRequest:
      type: object
      additionalProperties: false
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: The name of the todo item to create.
        list:
          $ref: "#/components/schemas/TodoListName"
          default: todo
        action_group:
          type:
            - string
            - "null"
          maxLength: 255
        action_date:
          type:
            - string
            - "null"
          format: date
        action:
          type:
            - string
            - "null"
          maxLength: 255
    TodoCreateResponse:
      type: object
      required:
        - success
        - item_id
        - title
        - list
      properties:
        success:
          type: boolean
        item_id:
          type: string
        title:
          type: string
        list:
          $ref: "#/components/schemas/TodoListName"
        action_group:
          type:
            - string
            - "null"
          maxLength: 255
        action_date:
          type:
            - string
            - "null"
          format: date
        action:
          type:
            - string
            - "null"
          maxLength: 255
    TodoUpdateActionMetadataRequest:
      type: object
      additionalProperties: false
      required:
        - list
      properties:
        list:
          $ref: "#/components/schemas/TodoListName"
        action_group:
          type:
            - string
            - "null"
          maxLength: 255
        action_date:
          type:
            - string
            - "null"
          format: date
        action:
          type:
            - string
            - "null"
          maxLength: 255
    TodoUpdateActionMetadataResponse:
      type: object
      required:
        - success
        - item_id
        - list
      properties:
        success:
          type: boolean
        item_id:
          type: string
        list:
          $ref: "#/components/schemas/TodoListName"
        action_group:
          type:
            - string
            - "null"
          maxLength: 255
        action_date:
          type:
            - string
            - "null"
          format: date
        action:
          type:
            - string
            - "null"
          maxLength: 255
    TodoBulkActionMetadataUpdate:
      type: object
      additionalProperties: false
      required:
        - item_id
        - list
      properties:
        item_id:
          type: string
          minLength: 1
          description: Monday.com item id to update.
        list:
          $ref: "#/components/schemas/TodoListName"
        action_group:
          type:
            - string
            - "null"
          maxLength: 255
        action_date:
          type:
            - string
            - "null"
          format: date
        action:
          type:
            - string
            - "null"
          maxLength: 255
    TodoBulkActionMetadataRequest:
      type: object
      additionalProperties: false
      required:
        - updates
      properties:
        updates:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/TodoBulkActionMetadataUpdate"
    TodoBulkActionMetadataResult:
      type: object
      required:
        - success
        - item_id
        - list
      properties:
        success:
          type: boolean
        item_id:
          type: string
        list:
          $ref: "#/components/schemas/TodoListName"
        action_group:
          type:
            - string
            - "null"
        action_date:
          type:
            - string
            - "null"
          format: date
        action:
          type:
            - string
            - "null"
        error:
          oneOf:
            - type: string
            - type: object
            - type: "null"
    TodoBulkActionMetadataResponse:
      type: object
      required:
        - success
        - updated_count
        - failed_count
        - results
      properties:
        success:
          type: boolean
          description: True only when every item in the batch updated successfully.
        updated_count:
          type: integer
        failed_count:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/TodoBulkActionMetadataResult"
    TodoItem:
      type: object
      required:
        - item_id
        - title
        - list
      properties:
        item_id:
          type: string
        title:
          type: string
        list:
          $ref: "#/components/schemas/TodoListName"
        group_id:
          type:
            - string
            - "null"
        group_title:
          type:
            - string
            - "null"
        action_group:
          type:
            - string
            - "null"
          maxLength: 255
        action_date:
          type:
            - string
            - "null"
          format: date
        action:
          type:
            - string
            - "null"
          maxLength: 255
    TodoListResponse:
      type: object
      required:
        - success
        - count
        - items
      properties:
        success:
          type: boolean
        count:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/TodoItem"
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer